dashboard/requirements.txt
streamlit==1.32.2
redis==5.0.1
pandas==2.1.4
streamlit-autorefresh==1.0.1
python-dotenv==1.0.1
